+++ sources/plugins/PhilipsHue/Entities/HueLightInformations.h
@@ -3,6 +3,9 @@
 #include "HueSwUpdate.h"
 #include "HueCapabilities.h"
 #include "HueConfig.h"
+#include "../Devices/Utils/LightModel.h"
+#include "../Devices/Utils/ColorType.h"
+#include "../Devices/Utils/LightType.h"
 
 class CHueLightInformations
 {
@@ -14,9 +17,9 @@ public:
 
    const CHueState& getState() const;
    const CHueSwUpdate& getSwUpdate() const;
-   const std::string& getType() const;
+   const ELightType& getType() const;
    const std::string& getName() const;
-   const std::string& getModelId() const;
+   const ELightModel& getModelId() const;
    const std::string& getManufacturerName() const;
    const std::string& getProductName() const;
    const CHueCapabilities& getCapabilities() const;
@@ -25,10 +28,11 @@ public:
    const std::string& getSwVersion() const;
    const std::string& getSwConfigId() const;
    const std::string& getProductId() const;
+   const EColorType& getColorType() const;
 
    void setState(const CHueState& state);
    void setSwUpdate(const CHueSwUpdate& swUpdate);
-   void setType(const std::string& type);
+   void setType(const ELightType& type);
    void setName(const std::string& name);
    void setModelId(const std::string& modelId);
    void setManufacturerName(const std::string& manufacturerName);
@@ -39,13 +43,14 @@ public:
    void setSwVersion(const std::string& swVersion);
    void setSwConfigId(const std::string& swConfigId);
    void setProductId(const std::string& productId);
+   void setColorType(const EColorType& colorType);
 
 private:
    CHueState m_state;
    CHueSwUpdate m_swUpdate;
-   std::string m_type;
+   ELightType m_type;
    std::string m_name;
-   std::string m_modelId;
+   ELightModel m_modelId;
    std::string m_manufacturerName;
    std::string m_productName;
    CHueCapabilities m_capabilities;
@@ -54,4 +59,5 @@ private:
    std::string m_swVersion;
    std::string m_swConfigId;
    std::string m_productId;
+   EColorType m_colorType;
 };