+++ sources/plugins/PhilipsHue/changelog.md
@@ -1,2 +1,7 @@
+### 1.1.1
+* [712](https://github.com/Yadoms/yadoms/issues/712) : Change friendly hue light name
+### 1.1.0
+* Historize light state when detecting / turning on/off a light #734
+* Add interfaces by device model #713
 ### 1.0.0
 * First release